refactor: extract sparse pattern handling out of porcelain
